+++ web/app/themes/lark-child/resources/views/components/blocks/container.blade.php
@@ -1,50 +1,49 @@
-@php
-if( isset($background_image) ) {
-  $srcset = wp_get_attachment_image_srcset($background_image, 'large_square');
-}
-@endphp
-<section
-  @if( isset($srcset) && ! empty($srcset) )
-    data-background-image-srcset="{{ $srcset }}"
-  @endif
+<section class="fcb
 
-  @if( isset($style) )
-    style="{{ $style }}"
-  @endif
-
-  @hassub('id')
-    id="{{ str_replace(' ', '-', preg_replace('/\s+/', ' ', strtolower(get_sub_field('id')))) }}"
-  @endsub
-
-  class="fcb
   @isset($classes)
     {{ $classes }}
   @endisset
 
-  @hassub('padding_override')
-    {{ 'fcb-' }}@sub('padding_override'){{ $defaultPadding }}
+  @hassub('background_colour')
+    bg-@sub('background_colour')
   @endsub
 
-  @hassub('background_colour')
-    {{ 'fcb-' }}@sub('background_colour')
-  @endsub
-">
+  @if ($overridePadding)
+    {{ $overridePadding }}
+  @else
+    @if (in_array(get_sub_field('background_colour'), ['default', 'white']))
+      {{-- TODO:  default padding for when the default colour is selected --}}
+    @else
+      {{ $defaultPadding }}
+    @endif
+  @endif
+
+  @if(get_sub_field('has_overlay') == true )
+    blend
+  @endif
+"
+@if( isset($style) )
+  style="{{ $style }}"
+@endif
+
+@hassub('background_image')
+  style="background-image: url('{{ wp_get_attachment_image_src(get_sub_field('background_image')['ID'], 'full')[0] }}');"
+@endsub
+>
   {{ $beforeSlot }}
 
-  @if ( ! isset($noContainer))
-    <div class="
-    @isset($overrideContainerClasses)
-      {{ $overrideContainerClasses}}
-    @else
-      container
-    @endisset
+  <div class="
+  @isset($overrideContainerClasses)
+    {{ $overrideContainerClasses}}
+  @else
+    container max-w-screen-xl
+  @endisset
 
-    @isset($containerClasses)
-      {{ $containerClasses}}
-    @endisset">
-      {!! $slot !!}
-    </div>
-  @endif
+  @isset($containerClasses)
+    {{ $containerClasses}}
+  @endisset">
+    {!! $slot !!}
+  </div>
 
   {{ $afterSlot }}
 </section>